script/requirements.txt
# 数据导入脚本依赖
mysql-connector-python==8.2.0
# 批量翻译脚本依赖
requests==2.31.0
# 中文分词和词性标注，用于智能翻译提取
jieba==0.42.1
# 用于处理JSON数据
# 用于日志记录
# 用于正则表达式处理
# 用于类型提示